+++ src/components/error.tsx
@@ -2,4 +2,4 @@ import React from 'react';
 
 export const Error = (): JSX.Element => {
   return <h2>404 not found</h2>
-}
+};
