+++ src/components/sandbox.tsx
@@ -1,15 +1,6 @@
 import React from 'react';
 import { MovieCard } from '../containers/movieCard';
 
-//export type MovieType = {
-//   Poster: string;
-//   Title: string;
-//   Type: string;
-//   Year: string;
-//   imdbID: string;
-// }
-
-
 export const Sandbox = (): JSX.Element => {
   return (
     <MovieCard
