+++ src/styles.css
@@ -135,7 +135,6 @@ h1 {
   }
 }
 
-
 @media screen and (max-width: 651px) {
   .single-movie-element {
     max-width: 100%;
